+++ src/include/common/file_utils.h
@@ -24,6 +24,8 @@ typedef enum PGFileType
 	PGFILETYPE_LNK
 } PGFileType;
 
+struct iovec;					/* avoid including port/pg_iovec.h here */
+
 #ifdef FRONTEND
 extern int	fsync_fname(const char *fname, bool isdir);
 extern void fsync_pgdata(const char *pg_data, int serverVersion);
@@ -36,5 +38,10 @@ extern PGFileType get_dirent_type(const char *path,
 								  const struct dirent *de,
 								  bool look_through_symlinks,
 								  int elevel);
+
+extern ssize_t pg_pwritev_with_retry(int fd,
+									 const struct iovec *iov,
+									 int iovcnt,
+									 off_t offset);
 
 #endif							/* FILE_UTILS_H */
